// packages/ashura-dialect-ui/src/language/generated/ast.ts
export interface Component extends langium.AstNode {
    readonly $container: Model;
    readonly $type: 'Component';
    name: string;
    trait: langium.Reference<Trait>;
    variables?: UiVariableBlock;
}

export const Component = {
    $type: 'Component',
    name: 'name',
    trait: 'trait',
    variables: 'variables'
} as const;

export function isComponent(item: unknown): item is Component {
    return reflection.isInstance(item, Component.$type);
}

export interface Model extends langium.AstNode {
    readonly $type: 'Model';
    elements: Array<UiElement>;
}

export const Model = {
    $type: 'Model',
    elements: 'elements'
} as const;

export function isModel(item: unknown): item is Model {
    return reflection.isInstance(item, Model.$type);
}

export interface Trait extends langium.AstNode {
    readonly $container: Model;
    readonly $type: 'Trait';
    members: Array<TraitMember>;
    name: string;
    states: Array<UiState>;
}

export const Trait = {
    $type: 'Trait',
    members: 'members',
    name: 'name',
    states: 'states'
} as const;

export function isTrait(item: unknown): item is Trait {
    return reflection.isInstance(item, Trait.$type);
}

export type TraitMember = UiForbidden | UiTransition;

export const TraitMember = {
    $type: 'TraitMember'
} as const;

export function isTraitMember(item: unknown): item is TraitMember {
    return reflection.isInstance(item, TraitMember.$type);
}

export type UiElement = Component | Trait;

export const UiElement = {
    $type: 'UiElement'
} as const;

export function isUiElement(item: unknown): item is UiElement {
    return reflection.isInstance(item, UiElement.$type);
}

export interface UiForbidden extends langium.AstNode {
    readonly $container: Trait;
    readonly $type: 'UiForbidden';
    from: langium.Reference<UiState>;
    to: langium.Reference<UiState>;
}

export const UiForbidden = {
    $type: 'UiForbidden',
    from: 'from',
    to: 'to'
} as const;
// ...
export interface UiState extends langium.AstNode {
    readonly $container: Trait;
    readonly $type: 'UiState';
    name: string;
}

export const UiState = {
    $type: 'UiState',
    name: 'name'
} as const;

export function isUiState(item: unknown): item is UiState {
    return reflection.isInstance(item, UiState.$type);
}

export interface UiTransition extends langium.AstNode {
    readonly $container: Trait;
    readonly $type: 'UiTransition';
    from: langium.Reference<UiState>;
    guard: UiGuardSpec;
    to: langium.Reference<UiState>;
}

export const UiTransition = {
    $type: 'UiTransition',
    from: 'from',
    guard: 'guard',
    to: 'to'
} as const;

export function isUiTransition(item: unknown): item is UiTransition {
    return reflection.isInstance(item, UiTransition.$type);
}

export interface UiVariableBlock extends langium.AstNode {
    readonly $container: Component;
    readonly $type: 'UiVariableBlock';
    entries: Array<UiVariableEntry>;
}

export const UiVariableBlock = {
    $type: 'UiVariableBlock',
    entries: 'entries'
} as const;

export function isUiVariableBlock(item: unknown): item is UiVariableBlock {
    return reflection.isInstance(item, UiVariableBlock.$type);
}

export interface UiVariableEntry extends langium.AstNode {
    readonly $container: UiVariableBlock;
    readonly $type: 'UiVariableEntry';
    name: string;
    type: string;
}

export const UiVariableEntry = {
    $type: 'UiVariableEntry',
    name: 'name',
    type: 'type'
} as const;